Use cell for workspace redirect lookups
