+++ front-api/routes/auth-context.ts
@@ -1,8 +1,7 @@
-import { getWorkspaceRegionRedirect } from "@app/lib/api/regions/lookup";
+import { getWorkspaceCellRedirect } from "@app/lib/api/cells/lookup";
 import { fetchUserFromSession } from "@app/lib/iam/users";
 import { sessionApp } from "@front-api/middlewares/ctx";
 import { apiError } from "@front-api/middlewares/utils";
-
 import { sessionAuth } from "../middlewares/session_auth";
 
 export const authContextApp = sessionApp();
@@ -13,13 +12,13 @@ authContextApp.get("/", async (ctx) => {
   const session = ctx.get("session");
 
   if (session.workspaceId) {
-    const redirect = await getWorkspaceRegionRedirect(session.workspaceId);
+    const redirect = await getWorkspaceCellRedirect(session.workspaceId);
     if (redirect) {
       return apiError(ctx, {
         status_code: 400,
         api_error: {
-          type: "workspace_in_different_region",
-          message: "Workspace is located in a different region",
+          type: "workspace_in_different_cell",
+          message: "Workspace is located in a different cell",
           redirect,
         },
       });
