+++ front/types/error.ts
@@ -2,10 +2,10 @@
 
 import { CONVERSATION_ERROR_TYPES } from "@app/types/assistant/conversation";
 import type { CoreAPIError } from "@app/types/core/core_api";
-import type { RegionType } from "@app/types/region";
 // biome-ignore lint/plugin/enforceClientTypesInPublicApi: existing usage
 import type { ConnectorsAPIError } from "@dust-tt/client";
 import type { ContentfulStatusCode } from "hono/utils/http-status";
+import type { CellInfo } from "./cell";
 
 const API_ERROR_TYPES = [
   "not_authenticated",
@@ -37,7 +37,8 @@ const API_ERROR_TYPES = [
   "provider_not_found",
   "dataset_not_found",
   "workspace_not_found",
-  "workspace_in_different_region",
+  "workspace_in_different_region", // TODO(single-tenant): remove this once we have fully migrated to cell.
+  "workspace_in_different_cell",
   "workspace_auth_error",
   "workspace_can_use_product_required_error",
   "workspace_user_not_found",
@@ -177,18 +178,16 @@ const API_ERROR_TYPES = [
   "wakeup_not_found",
 ] as const;
 
-export type RegionRedirectError = {
-  region: RegionType;
-  url: string;
-};
+export type RegionRedirectError = CellInfo;
 
 export type APIErrorType = (typeof API_ERROR_TYPES)[number];
 
 // Error types that are expected outcomes of normal operation rather than
-// failures (e.g. a region redirect). Callers can use this to log them at a
+// failures (e.g. a cell redirect). Callers can use this to log them at a
 // lower level so they don't pollute error monitoring.
 export const EXPECTED_API_ERROR_TYPES: ReadonlySet<APIErrorType> = new Set([
-  "workspace_in_different_region",
+  "workspace_in_different_region", // TODO(single-tenant): remove this once we have fully migrated to cell.
+  "workspace_in_different_cell",
 ]);
 
 export type APIError = {
@@ -198,7 +197,7 @@ export type APIError = {
   run_error?: CoreAPIError;
   app_error?: CoreAPIError;
   connectors_error?: ConnectorsAPIError;
-  redirect?: RegionRedirectError;
+  redirect?: CellInfo;
   unverifiableRefs?: string[];
 };
