+++ front/types/shared/cache_resource_registry.ts
@@ -114,21 +114,6 @@ export const CACHE_RESOURCE_REGISTRY: CacheResourceDefinition[] = [
     resolverKeyPattern: "workos-orgs-*",
   },
   {
-    id: "workspace_region",
-    label: "Workspace region",
-    fnName: "_lookupWorkspaceUncached",
-    params: [
-      {
-        key: "wId",
-        label: "Workspace sId",
-        type: "string",
-        placeholder: "e.g. abc123",
-      },
-    ],
-    buildResolverKey: (p) => `workspace-region:${p.wId}`,
-    resolverKeyPattern: "workspace-region:*",
-  },
-  {
     id: "workspace_cell",
     label: "Workspace cell",
     fnName: "_lookupWorkspaceUncached",
